+++ service/controller/clusternamespace/clusternamespace.go
@@ -1,71 +1,71 @@
-package orgcluster
+package clusternamespace
 
 import (
 	// If your operator watches a CRD import it here.
 	// "github.com/giantswarm/apiextensions/v2/pkg/apis/application/v1alpha1"
 
-	"github.com/giantswarm/apiextensions/v3/pkg/label"
 	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"github.com/giantswarm/operatorkit/v4/pkg/controller"
 	"github.com/giantswarm/operatorkit/v4/pkg/resource"
+	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
-	cluster "sigs.k8s.io/cluster-api/api/v1alpha3"
 
+	"github.com/giantswarm/rbac-operator/pkg/label"
 	"github.com/giantswarm/rbac-operator/pkg/project"
 )
 
-type OrgClusterConfig struct {
+type ClusterNamespaceConfig struct {
 	K8sClient k8sclient.Interface
 	Logger    micrologger.Logger
 }
-type OrgCluster struct {
+type ClusterNamespace struct {
 	*controller.Controller
 }
 
-func NewOrgCluster(config OrgClusterConfig) (*OrgCluster, error) {
+func NewClusterNamespace(config ClusterNamespaceConfig) (*ClusterNamespace, error) {
 	var err error
 
 	var resources []resource.Interface
 	{
-		c := orgClusterResourcesConfig(config)
+		c := clusterNamespaceResourcesConfig(config)
 
-		resources, err = newOrgClusterResources(c)
+		resources, err = newClusterNamespaceResources(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
 	}
 
-	var orgClusterController *controller.Controller
+	var clusterNamespaceController *controller.Controller
 	{
 		selector := newSelector(func(labels labels.Labels) bool {
-			return labels.Has(label.Organization)
+			return labels.Has(label.Organization) && labels.Has(label.Cluster)
 		})
 
 		c := controller.Config{
 			K8sClient: config.K8sClient,
 			Logger:    config.Logger,
 			NewRuntimeObjectFunc: func() runtime.Object {
-				return new(cluster.Cluster)
+				return new(corev1.Namespace)
 			},
 			Resources: resources,
 			Selector:  selector,
 
 			// Name is used to compute finalizer names. This here results in something
-			// like operatorkit.giantswarm.io/rbac-operator-org-cluster-controller.
-			Name: project.Name() + "-org-cluster-controller",
+			// like operatorkit.giantswarm.io/rbac-operator-cluster-namespace-controller.
+			Name: project.Name() + "-cluster-namespace-controller",
 		}
 
-		orgClusterController, err = controller.New(c)
+		clusterNamespaceController, err = controller.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
 	}
 
-	c := &OrgCluster{
-		Controller: orgClusterController,
+	c := &ClusterNamespace{
+		Controller: clusterNamespaceController,
 	}
 
 	return c, nil