+++ service/service.go
@@ -6,6 +6,7 @@ import (
 	"context"
 	"sync"
 
+	securityv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/security/v1alpha1"
 	"github.com/giantswarm/k8sclient/v5/pkg/k8sclient"
 	"github.com/giantswarm/k8sclient/v5/pkg/k8srestconfig"
 	"github.com/giantswarm/microendpoint/service/version"
@@ -13,12 +14,11 @@ import (
 	"github.com/giantswarm/micrologger"
 	"github.com/spf13/viper"
 	"k8s.io/client-go/rest"
-	cluster "sigs.k8s.io/cluster-api/api/v1alpha3"
 
 	"github.com/giantswarm/rbac-operator/flag"
 	"github.com/giantswarm/rbac-operator/pkg/project"
 	"github.com/giantswarm/rbac-operator/service/collector"
-	"github.com/giantswarm/rbac-operator/service/controller/orgcluster"
+	"github.com/giantswarm/rbac-operator/service/controller/clusternamespace"
 	"github.com/giantswarm/rbac-operator/service/controller/orgpermissions"
 	"github.com/giantswarm/rbac-operator/service/controller/rbac"
 	"github.com/giantswarm/rbac-operator/service/internal/bootstrap"
@@ -35,12 +35,12 @@ type Config struct {
 type Service struct {
 	Version *version.Service
 
-	bootOnce                 sync.Once
-	bootstrapRunner          *bootstrap.Bootstrap
-	rbacController           *rbac.RBAC
-	orgPermissionsController *orgpermissions.OrgPermissions
-	orgClusterController     *orgcluster.OrgCluster
-	operatorCollector        *collector.Set
+	bootOnce                   sync.Once
+	bootstrapRunner            *bootstrap.Bootstrap
+	rbacController             *rbac.RBAC
+	orgPermissionsController   *orgpermissions.OrgPermissions
+	clusterNamespaceController *clusternamespace.ClusterNamespace
+	operatorCollector          *collector.Set
 }
 
 // New creates a new configured service object.
@@ -92,7 +92,7 @@ func New(config Config) (*Service, error) {
 		c := k8sclient.ClientsConfig{
 			Logger: config.Logger,
 			SchemeBuilder: k8sclient.SchemeBuilder{
-				cluster.AddToScheme,
+				securityv1alpha1.AddToScheme,
 			},
 			RestConfig: restConfig,
 		}
@@ -119,15 +119,15 @@ func New(config Config) (*Service, error) {
 		}
 	}
 
-	var orgClusterController *orgcluster.OrgCluster
+	var clusterNamespaceController *clusternamespace.ClusterNamespace
 	{
 
-		c := orgcluster.OrgClusterConfig{
+		c := clusternamespace.ClusterNamespaceConfig{
 			K8sClient: k8sClient,
 			Logger:    config.Logger,
 		}
 
-		orgClusterController, err = orgcluster.NewOrgCluster(c)
+		clusterNamespaceController, err = clusternamespace.NewClusterNamespace(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -195,12 +195,12 @@ func New(config Config) (*Service, error) {
 	s := &Service{
 		Version: versionService,
 
-		bootOnce:                 sync.Once{},
-		bootstrapRunner:          bootstrapRunner,
-		rbacController:           rbacController,
-		orgPermissionsController: orgPermissionsController,
-		orgClusterController:     orgClusterController,
-		operatorCollector:        operatorCollector,
+		bootOnce:                   sync.Once{},
+		bootstrapRunner:            bootstrapRunner,
+		rbacController:             rbacController,
+		orgPermissionsController:   orgPermissionsController,
+		clusterNamespaceController: clusterNamespaceController,
+		operatorCollector:          operatorCollector,
 	}
 
 	return s, nil
@@ -225,6 +225,6 @@ func (s *Service) Boot(ctx context.Context) {
 
 		go s.orgPermissionsController.Boot(ctx)
 
-		go s.orgClusterController.Boot(ctx)
+		go s.clusterNamespaceController.Boot(ctx)
 	})
 }