+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add org-cluster controller which ensures that RBAC resources to access resources in cluster namespaces can be granted to those with access to the clusters organization
+- Add cluster-namespace controller which ensures that RBAC resources to access resources in cluster namespaces can be granted to those with access to the clusters organization
 - Add bootstrapping for the `read-cluster-apps` and `write-cluster-apps` clusterRoles.
 
 ## [0.20.0] - 2022-02-07